+++ source/ci_source/providers/BitbucketPipelines.ts
@@ -26,7 +26,6 @@ import { ensureEnvKeysExist, ensureEnvKeysAreInt } from "../ci_source_helpers"
  *
  * ### Token Setup
  *
- * Add `DANGER_BITBUCKETCLOUD_UUID` to your pipeline repository variable.
  * You can either add `DANGER_BITBUCKETCLOUD_USERNAME`, `DANGER_BITBUCKETCLOUD_PASSWORD`
  * or add `DANGER_BITBUCKETCLOUD_OAUTH_KEY`, `DANGER_BITBUCKETCLOUD_OAUTH_SECRET`
  * -