+++ source/runner/Executor.ts
@@ -294,7 +294,7 @@ export class Executor {
         let comment
         if (process.env["DANGER_BITBUCKETSERVER_HOST"]) {
           comment = bitbucketServerTemplate(dangerID, mergedResults, commitID)
-        } else if (process.env["DANGER_BITBUCKETCLOUD_UUID"]) {
+        } else if (process.env["DANGER_BITBUCKETCLOUD_OAUTH_KEY"] || process.env["DANGER_BITBUCKETCLOUD_USERNAME"]) {
           comment = bitbucketCloudTemplate(dangerID, mergedResults, commitID)
         } else {
           comment = githubResultsTemplate(dangerID, mergedResults, commitID)
@@ -414,7 +414,7 @@ export class Executor {
     let comment
     if (process.env["DANGER_BITBUCKETSERVER_HOST"]) {
       comment = bitbucketServerInlineTemplate(this.options.dangerID, results, inlineResults.file, inlineResults.line)
-    } else if (process.env["DANGER_BITBUCKETCLOUD_UUID"]) {
+    } else if (process.env["DANGER_BITBUCKETCLOUD_OAUTH_KEY"] || process.env["DANGER_BITBUCKETCLOUD_USERNAME"]) {
       comment = bitbucketCloudInlineTemplate(this.options.dangerID, results, inlineResults.file, inlineResults.line)
     } else {
       comment = githubResultsInlineTemplate(this.options.dangerID, results, inlineResults.file, inlineResults.line)
@@ -445,7 +445,7 @@ const messageForResults = (results: DangerResults) => {
   } else {
     if (process.env["DANGER_BITBUCKETSERVER_HOST"]) {
       return bitbucketMessageForResultWithIssues
-    } else if (process.env["DANGER_BITBUCKETCLOUD_UUID"]) {
+    } else if (process.env["DANGER_BITBUCKETCLOUD_OAUTH_KEY"] || process.env["DANGER_BITBUCKETCLOUD_USERNAME"]) {
       return bitbucketCloudMessageForResultWithIssues
     } else {
       return githubMessageForResultWithIssues