+++ CHANGELOG.md
@@ -15,7 +15,8 @@ x
 
 <!-- Your comment below this -->
 
-<!-- Your comment above this -->
+- Bitbucket Cloud: Allow DangerCI to get UUID from Bitbucket - [@hellocore]
+  <!-- Your comment above this -->
 
 # 9.2.7-8-9